+++ src/controllers/utils/instances.py
@@ -3,7 +3,7 @@ from dotenv import load_dotenv
 from src.controllers.smart_plug_controller import SmartPlugController
 from src.controllers.spotify_controller import SpotifyController
 from src.controllers.tv_controller import TVController
-from src.gpio_setup import instantiate_button_controller
+from src.controllers.utils.gpio_setup import instantiate_button_controller
 from src.utils.counter import PlaybackCounter
 
 load_dotenv()
